move syntax::parse::lexer::comments -> syntax::util::comments
